requirements.txt
urllib3==2.0.2
unstructured==0.6.6
extract-msg==0.41.1
tabulate==0.9.0
python-multipart==0.0.6
fastapi==0.96.0
SQLAlchemy==2.0.15
alembic==1.11.1
requests==2.31.0
PyYAML==6.0.1
psutil==5.9.6
python-on-whales==0.65.0